+++ src/cecj/eval/SimpleCoevolutionaryEvaluator.java
@@ -149,7 +149,7 @@ public class SimpleCoevolutionaryEvaluator extends CoevolutionaryEvaluator {
 		evaluationsPerformed += interactionScheme.getEvaluationsNumber(state, opponents, true);
 
 		for (int subpop = 0; subpop < numSubpopulations; subpop++) {
-			int[][] subpopulationResults = interactionScheme.performInteractions(state, subpop,
+			float[][] subpopulationResults = interactionScheme.performInteractions(state, subpop,
 					opponents);
 
 			fitnessAggregateMethod[subpop].prepareToAggregate(state, subpop);
